Create SalaryPointer.c
Programming Task
Create a program to change the value of a variable using a pointer. Here's how the program should work:

Get input value for a double variable salary.
Assign the address of salary to a double pointer.
Now use the pointer to print the value of salary, increase the salary by 2 times, print the new salary.